Move code to src and separate blocks
